test: Move the review-fix tests into the files they belong with
review_fixes_test.go was named for when it was written rather than what
it covers, which would mean nothing in six months. The annotation value
and bulk tests go with the other annotation tests, the catch-all auth
tests with the middleware tests, and the discovered_at polling tests
with the handler tests.

No behavior or assertions changed.
